+++ classes/Email.php
@@ -2,11 +2,11 @@
 namespace Grav\Plugin\Email;
 
 use Grav\Common\Config\Config;
-use Grav\Common\Filesystem\Folder;
 use Grav\Common\Grav;
+use Grav\Common\Language\Language;
+use Grav\Common\Twig\Twig;
 use \Monolog\Logger;
 use \Monolog\Handler\StreamHandler;
-use RocketTheme\Toolbox\ResourceLocator\UniformResourceLocator;
 
 class Email
 {
@@ -118,6 +118,218 @@ class Email
         }
 
         return $result;
+    }
+
+    /**
+     * Build e-mail message.
+     *
+     * @param array $params
+     * @param array $vars
+     * @return \Swift_Message
+     */
+    public function buildMessage(array $params, array $vars = array())
+    {
+        /** @var Twig $twig */
+        $twig = Grav::instance()['twig'];
+
+        /** @var Config $config */
+        $config = Grav::instance()['config'];
+
+        /** @var Language $language */
+        $language = Grav::instance()['language'];
+
+        // Extend parameters with defaults.
+        $params += array(
+            'bcc' => $config->get('plugins.email.bcc', array()),
+            'body' => $config->get('plugins.email.body', '{% include "forms/data.html.twig" %}'),
+            'cc' => $config->get('plugins.email.cc', array()),
+            'cc_name' => $config->get('plugins.email.cc_name'),
+            'charset' =>  $config->get('plugins.email.charset', 'utf-8'),
+            'from' => $config->get('plugins.email.from'),
+            'from_name' => $config->get('plugins.email.from_name'),
+            'content_type' => $config->get('plugins.email.content_type', 'text/html'),
+            'reply_to' => $config->get('plugins.email.reply_to', array()),
+            'reply_to_name' => $config->get('plugins.email.reply_to_name'),
+            'subject' => !empty($vars['form']) && $vars['form'] instanceof Form ? $vars['form']->page()->title() : null,
+            'to' => $config->get('plugins.email.to'),
+            'to_name' => $config->get('plugins.email.to_name'),
+            'process_markdown' => false,
+            'template' => false
+        );
+
+        // Create message object.
+        $message = $this->message();
+
+        if (!$params['to']) {
+            throw new \RuntimeException($language->translate('PLUGIN_EMAIL.PLEASE_CONFIGURE_A_TO_ADDRESS'));
+        }
+        if (!$params['from']) {
+            throw new \RuntimeException($language->translate('PLUGIN_EMAIL.PLEASE_CONFIGURE_A_FROM_ADDRESS'));
+        }
+
+        // Process parameters.
+        foreach ($params as $key => $value) {
+            switch ($key) {
+                case 'body':
+                    if (is_string($value)) {
+                        $body = $twig->processString($value, $vars);
+
+                        if ($params['process_markdown']) {
+                            $parsedown = new \Parsedown();
+                            $body = $parsedown->text($body);
+                        }
+
+                        if ($params['template']) {
+                            $body = $twig->processTemplate($params['template'], ['content' => $body]);
+                        }
+
+                        $content_type = !empty($params['content_type']) ? $twig->processString($params['content_type'], $vars) : null;
+                        $charset = !empty($params['charset']) ? $twig->processString($params['charset'], $vars) : null;
+
+                        $message->setBody($body, $content_type, $charset);
+                    }
+                    elseif (is_array($value)) {
+                        foreach ($value as $body_part) {
+                            $body_part += array(
+                                'charset' => $params['charset'],
+                                'content_type' => $params['content_type'],
+                            );
+
+                            $body = !empty($body_part['body']) ? $twig->processString($body_part['body'], $vars) : null;
+
+                            if ($params['process_markdown']) {
+                                $parsedown = new \Parsedown();
+                                $body = $parsedown->text($body);
+                            }
+
+                            $content_type = !empty($body_part['content_type']) ? $twig->processString($body_part['content_type'], $vars) : null;
+                            $charset = !empty($body_part['charset']) ? $twig->processString($body_part['charset'], $vars) : null;
+
+                            if (!$message->getBody()) {
+                                $message->setBody($body, $content_type, $charset);
+                            }
+                            else {
+                                $message->addPart($body, $content_type, $charset);
+                            }
+                        }
+                    }
+                    break;
+
+                case 'subject':
+                    $message->setSubject($twig->processString($language->translate($value), $vars));
+                    break;
+
+                case 'to':
+                    if (is_string($value) && !empty($params['to_name'])) {
+                        $value = array(
+                            'mail' => $twig->processString($value, $vars),
+                            'name' => $twig->processString($params['to_name'], $vars),
+                        );
+                    }
+
+                    foreach ($this->parseAddressValue($value, $vars) as $address) {
+                        $message->addTo($address->mail, $address->name);
+                    }
+                    break;
+
+                case 'cc':
+                    if (is_string($value) && !empty($params['cc_name'])) {
+                        $value = array(
+                            'mail' => $twig->processString($value, $vars),
+                            'name' => $twig->processString($params['cc_name'], $vars),
+                        );
+                    }
+
+                    foreach ($this->parseAddressValue($value, $vars) as $address) {
+                        $message->addCc($address->mail, $address->name);
+                    }
+                    break;
+
+                case 'bcc':
+                    foreach ($this->parseAddressValue($value, $vars) as $address) {
+                        $message->addBcc($address->mail, $address->name);
+                    }
+                    break;
+
+                case 'from':
+                    if (is_string($value) && !empty($params['from_name'])) {
+                        $value = array(
+                            'mail' => $twig->processString($value, $vars),
+                            'name' => $twig->processString($params['from_name'], $vars),
+                        );
+                    }
+
+                    foreach ($this->parseAddressValue($value, $vars) as $address) {
+                        $message->addFrom($address->mail, $address->name);
+                    }
+                    break;
+
+                case 'reply_to':
+                    if (is_string($value) && !empty($params['reply_to_name'])) {
+                        $value = array(
+                            'mail' => $twig->processString($value, $vars),
+                            'name' => $twig->processString($params['reply_to_name'], $vars),
+                        );
+                    }
+
+                    foreach ($this->parseAddressValue($value, $vars) as $address) {
+                        $message->addReplyTo($address->mail, $address->name);
+                    }
+                    break;
+
+            }
+        }
+
+        return $message;
+    }
+
+    /**
+     * Return parsed e-mail address value.
+     *
+     * @param $value
+     * @param array $vars
+     * @return array
+     */
+    public function parseAddressValue($value, array $vars = array())
+    {
+        $parsed = array();
+
+        /** @var Twig $twig */
+        $twig = Grav::instance()['twig'];
+
+        // Single e-mail address string
+        if (is_string($value)) {
+            $parsed[] = (object) array(
+                'mail' => $twig->processString($value, $vars),
+                'name' => null,
+            );
+        }
+
+        else {
+            // Cast value as array
+            $value = (array) $value;
+
+            // Single e-mail address array
+            if (!empty($value['mail'])) {
+                $parsed[] = (object) array(
+                    'mail' => $twig->processString($value['mail'], $vars),
+                    'name' => !empty($value['name']) ? $twig->processString($value['name'], $vars) : NULL,
+                );
+            }
+
+            // Multiple addresses (either as strings or arrays)
+            elseif (!(empty($value['mail']) && !empty($value['name']))) {
+                foreach ($value as $y => $itemx) {
+                    $addresses = $this->parseAddressValue($itemx, $vars);
+
+                    if (($address = reset($addresses))) {
+                        $parsed[] = $address;
+                    }
+                }
+            }
+        }
+
+        return $parsed;
     }
 
     /**